fix: don't pull the Client Routing runtime into Server Routing apps upon importing vike/getPageContext (fix #3471)
The browser entry of vike/getPageContext imported getPageContextClient()
from renderPageClient.ts, which loads the entire Client Routing runtime —
including shared-server-client/route/index.js whose module-level
assertClientRouting() contradicts the Server Routing entry's
assertServerRouting(). Any Server Routing app whose browser bundle imports
vike/getPageContext (e.g. every vike-react app: onRenderClient ->
callCumulativeHooks -> providePageContext) therefore shipped ~28 kB of
unused Client Routing runtime and logged this warning on every page load:

  [vike][Warning] You shouldn't `import { something } from
  'vike/client/router'` when using Server Routing.

Fix: move getPageContextClient() and its pageContext slot into a new
dependency-light module getPageContextClient.ts (same pattern as the
existing getPageContextCurrent.ts). Its closure — getGlobalObject,
getPageContextPublicClient, assertEnvClient — never reaches route/index.js
nor assertRoutingType.js. Pure code motion; no behavior change:
renderPageClient.ts sets the slot at the same lifecycle point as before,
and vike/client/router still exports getPageContextClient.

Verified:
- Server Routing + vike-react repro: warning gone, client bundle -28.8 kB
  (10.9%), page renders and hydrates.
- Client Routing unaffected: test/vike-react, examples/react-full, and
  examples/react-minimal e2e all pass (dev + preview); unit tests 206/206.
